+++ tests/integration/chunkstore/test_chunkstore.py
@@ -1222,3 +1222,62 @@ def test_stats(chunkstore_lib):
     assert(s['symbols']['count'] == 5)
     assert(s['chunks']['count'] == 366 * 5)
     assert(s['chunks']['count'] == s['metadata']['count'])
+
+
+def test_metadata(chunkstore_lib):
+     df = DataFrame(data={'data': np.random.randint(0, 100, size=2)},
+                   index=pd.date_range('2016-01-01', '2016-01-02'))
+     df.index.name = 'date'
+     chunkstore_lib.write('data', df, metadata = 'some metadata')
+     m = chunkstore_lib.read_metadata('data')
+     assert(m == u'some metadata')
+
+
+def test_metadata_update(chunkstore_lib):
+    df = DataFrame(data={'data': np.random.randint(0, 100, size=2)},
+                   index=pd.date_range('2016-01-01', '2016-01-02'))
+    df.index.name = 'date'
+    chunkstore_lib.write('data', df, metadata = 'some metadata', chunk_size='M')
+   
+    df = DataFrame(data={'data': np.random.randint(0, 100, size=1)},
+                   index=pd.date_range('2016-01-02', '2016-01-02'))
+    df.index.name = 'date'
+    chunkstore_lib.update('data', df, metadata='different metadata')
+    m = chunkstore_lib.read_metadata('data')
+    assert(m == u'different metadata') 
+
+
+def test_metadata_nosymbol(chunkstore_lib):
+    with pytest.raises(NoDataFoundException):
+        chunkstore_lib.read_metadata('None')
+
+
+def test_metadata_none(chunkstore_lib):
+    df = DataFrame(data={'data': np.random.randint(0, 100, size=2)},
+                   index=pd.date_range('2016-01-01', '2016-01-02'))
+    df.index.name = 'date'
+    chunkstore_lib.write('data', df, chunk_size='M')
+    assert(chunkstore_lib.read_metadata('data') == None)
+
+
+def test_metadata_invalid(chunkstore_lib):
+    df = DataFrame(data={'data': np.random.randint(0, 100, size=2)},
+                   index=pd.date_range('2016-01-01', '2016-01-02'))
+    df.index.name = 'date'
+    with pytest.raises(Exception) as e:
+        chunkstore_lib.write('data', df, chunk_size='M', metadata=df)
+
+
+def test_write_metadata(chunkstore_lib):
+    df = DataFrame(data={'data': np.random.randint(0, 100, size=2)},
+                   index=pd.date_range('2016-01-01', '2016-01-02'))
+    df.index.name = 'date'
+    chunkstore_lib.write('data', df)
+    chunkstore_lib.write_metadata('data', 'meta')
+    m = chunkstore_lib.read_metadata('data')
+    assert(m == u'meta')
+
+
+def test_write_metadata_nosymbol(chunkstore_lib):
+    with pytest.raises(NoDataFoundException):
+        chunkstore_lib.write_metadata('doesnt_exist', 'meta')